requirements.txt
Django>=5.0,<6.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.2.0
psycopg2-binary>=2.9.9
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
pyotp>=2.9.0
qrcode>=7.4.2
pillow>=10.1.0
django-cors-headers>=4.3.1
django-recaptcha>=4.0.0
openpyxl>=3.0.0
reportlab>=4.0.9
django-anymail>=10.0
python-decouple>=3.8
celery>=5.3.0
redis>=5.0.0
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
django-filter>=23.5
channels>=4.0.0
channels-redis>=4.1.0
django-rosetta>=0.9.9
pycountry>=23.12.11
